+++ shared-libs/crates/start-core/src/util/mod.rs
@@ -33,7 +33,9 @@ use crate::util::serde::{deserialize_from_str, serialize_display};
 use crate::{Error, ErrorKind, ResultExt as _};
 
 pub mod actor;
+pub mod btrfs;
 pub mod clap;
+pub mod clone;
 pub mod collections;
 pub mod cpupower;
 pub mod crypto;